master can run and accept registrations from local daemons. master can ask local daemons for reports for given TaskID. reports still make it to final destination

+++ src/java/src/main/java/edu/berkeley/xtrace/localdaemon/MasterServiceImpl.java
@@ -5,13 +5,13 @@ import java.util.Collection;
 
 public class MasterServiceImpl implements MasterService.Iface {
 
-    public Collection<Short> daemons;
+    public Collection<Integer> daemons;
 
-    public MasterServiceImpl(Collection<Short> daemons) {
+    public MasterServiceImpl(Collection<Integer> daemons) {
         this.daemons = daemons;
     }
 
-    public void registerDaemon(long id, short port) throws TException {
+    public void registerDaemon(long id, int port) throws TException {
         System.out.println("Register Daemon: (" + id + ", " + port + ")");
         daemons.add(port);
     }
