+++ src/thrift/master.thrift
@@ -4,5 +4,5 @@ namespace java edu.berkeley.xtrace.localdaemon
 
 
 service MasterService { 
-        void registerDaemon(1:i64 id, 2: i16 port);
+        void registerDaemon(1:i64 id, 2: i32 port);
 }
